Refactoring: inline environment variable creation into `run_command` method
Take care that both `on_change` and `run_command` method signature has evolved by taking an `event` object as an optional argument.
The backward compatibility is ensured as the argument has a default value set to `None`.

Getting the `event` objet in the `run_command` method permits to get rid of the regular expression used in the log message.

Signed-off-by: Eric Villard <dev@eviweb.fr>